requirements.txt
pillow
playwright